seperate helpers for libvirt
