+++ tasks/kernel_matrix_testing/stacks.py
@@ -1,4 +1,5 @@
 from .init_kmt import KMT_STACKS_DIR, VMCONFIG, check_and_get_stack
+from .libvirt import delete_domains, delete_volumes, delete_pools, delete_networks
 import json
 import os
 from .tool import Exit, info, ask, warn, error
@@ -84,77 +85,6 @@ def launch_stack(ctx, stack, branch, ssh_key, x86_ami, arm_ami):
     ctx.run(
         f"{env_vars} {prefix} inv -e system-probe.start-microvms --instance-type-x86={X86_INSTANCE_TYPE} --instance-type-arm={ARM_INSTANCE_TYPE} --x86-ami-id={x86_ami} --arm-ami-id={arm_ami} --ssh-key-name={ssh_key} --infra-env=aws/sandbox --vmconfig={vm_config} --stack-name={stack}"
     )
-
-
-def resource_in_stack(stack, resource):
-    return stack in resource
-
-
-def get_resources_in_stack(stack, list_fn):
-    resources = list_fn()
-    stack_resources = list()
-    for resource in resources:
-        if resource_in_stack(stack, resource.name()):
-            stack_resources.append(resource)
-
-    return stack_resources
-
-
-def delete_domains(conn, stack):
-    domains = get_resources_in_stack(stack, conn.listAllDomains)
-    info(f"[*] {len(domains)} VMs running in stack {stack}")
-
-    for domain in domains:
-        name = domain.name()
-        domain.destroy()
-        domain.undefine()
-        info(f"[+] VM {name} deleted")
-
-
-def getAllStackVolumesFn(conn, stack):
-    def getAllStackVolumes():
-        pools = get_resources_in_stack(stack, conn.listAllStoragePools)
-
-        volumes = list()
-        for pool in pools:
-            volumes += pool.listAllVolumes()
-
-        return volumes
-
-    return getAllStackVolumes
-
-
-def delete_volumes(conn, stack):
-    volumes = get_resources_in_stack(stack, getAllStackVolumesFn(conn, stack))
-    info(f"[*] {len(volumes)} storage volumes running in stack {stack}")
-
-    for volume in volumes:
-        name = volume.name()
-        volume.delete()
-        #        volume.undefine()
-        info(f"[+] Storage volume {name} deleted")
-
-
-def delete_pools(conn, stack):
-    pools = get_resources_in_stack(stack, conn.listAllStoragePools)
-    info(f"[*] {len(pools)} storage pools running in stack {stack}")
-
-    for pool in pools:
-        name = pool.name()
-        pool.destroy()
-        pool.undefine()
-        info(f"[+] Storage pool {name} deleted")
-
-
-def delete_networks(conn, stack):
-    networks = get_resources_in_stack(stack, conn.listAllNetworks)
-    info(f"[*] {len(networks)} networks running in stack {stack}")
-
-    for network in networks:
-        name = network.name()
-        network.destroy()
-        network.undefine()
-        info(f"[+] Network {name} deleted")
 
 
 def destroy_stack_pulumi(ctx, stack, ssh_key):
